+++ src/models.ts
@@ -2,5 +2,5 @@ export { AcEntity } from './models/ac-entity';
 export { AcNotification } from './models/ac-notification';
 export { ActionType } from './models/action-type.enum';
 export { MapLayerProviderOptions } from './models/map-layer-provider-options.enum';
-export { BasicEntityDrawerOptions } from './models/basic-entity-drawer-options';
+export { EntitiesDrawerOptions } from './models/entities-drawer-options';
 export { LayerOptions } from './models/layer-options';