+++ src/models/entities-drawer-options.ts
@@ -1,4 +1,4 @@
-export interface BasicEntityDrawerOptions {
+export interface EntitiesDrawerOptions {
   collectionSuspensionTime: number;
   collectionsNumber: number;
   collectionMaxSize: number;