+++ src/services/drawers/polygon-drawer/polygon-drawer.service.ts
@@ -3,6 +3,9 @@ import { EntitiesDrawerService } from '../entities-drawer/entities-drawer.servic
 import { CesiumService } from '../../cesium/cesium.service';
 import { GraphicsType } from '../entities-drawer/enums/graphics-type.enum';
 
+/**
+ *  This drawer is responsible for drawing polygons.
+ */
 @Injectable()
 export class PolygonDrawerService extends EntitiesDrawerService {
     constructor(cesiumService: CesiumService) {